+++ kutana/storages/__init__.py
@@ -1,0 +1,2 @@
+from .memory import MemoryStorage
+from .mongodb import MongoDBStorage